src/research_first_step/requirements.txt
google-genai>=1.0.0
groq>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.0

